requirements.txt
fastapi
pydantic
pymongo
bcrypt
pydantic[email]
uvicorn
python-dotenv
retrying